Evaluate pending functions on every call to handle forward references with subscripts.
The logic in the `__call__` path has always been minimal, but there doesn't appear to be any better way to fully support subscripts. Calling `get_type` would be significantly slower, whereas a pending check is only ~6% slower in the worst case - a single arg builtin function. So in practice this performance hit should be negligible; speed is still comparable to `singledispatch`. Fixes #43.