Adaptive type checking of arguments.
The global `get_type` function is replaced with an adaptive version that's specific to each `subtype`.
Because Unions need to be supported anyway, a Union is used to repsent the set of subscripted types that must be checked.
This is of course slower, but only proportionally to how many subscripted types are in a given param.

Fixes #46.