Support types which can't be subclassed.
Types with subscripts subclass the origin type to support `issubclass` as the first argument. If the type is not an acceptable base type, then there is no general way to support it.

As a workaround, this fix falls back to subclassing the registered origin instead, which then requires filtering the results of a union type checker. Refs #84.